[Platinum III] Title: 회사 문화 2, Time: 60 ms, Memory: 13372 KB -BaekjoonHub

+++ 백준/Platinum/14268. 회사 문화 2/회사 문화 2.cc
@@ -19,7 +19,7 @@ https://www.acmicpc.net/problem/10999
 
 struct SegmentTree {
 
-    using DataType = long long;
+    using DataType = int;
     vector<DataType> tree;
     vector<DataType> lazy;
